refactor: 💡 useTimeout hook을 분리
특정 도메인에 종속적이지 않고, 추후 재사용성이 있는 hook이라 판단해 분리함.
