+++ frontend/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useTimeout } from './useTimeout';
